% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{calc.breaks.acut}
\alias{calc.breaks.acut}
\title{Calculate the breaks from a series of clusters}
\usage{
calc.breaks.acut(acut, start = 0, end = 1)
}
\value{
breaks
}
\description{
Calculate the breaks from a series of clusters
}
