% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{calc.breaks}
\alias{calc.breaks}
\title{Calculate the breaks from the cluster object}
\usage{
calc.breaks(ht, nclust, cocl)
}
\arguments{
\item{ht}{cluster object}

\item{nclust}{number of clusters}

\item{cocl}{clusters}
}
\value{
breaks
}
\description{
Calculate the breaks from the cluster object
}
