% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{calc.breaks.rect}
\alias{calc.breaks.rect}
\title{Calculate the breaks and rectangles from two clusters}
\usage{
calc.breaks.rect(hcls, vcls, colset, start = 0, end = 1)
}
\arguments{
\item{hcls}{horizontal clusters}

\item{vcls}{vertical clusters}

\item{colset}{colors for the clusters}
}
\value{
rectangles and colors
}
\description{
Calculate the breaks and rectangles from two clusters
}
