% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{clamp.mat}
\alias{clamp.mat}
\title{Clamp a matrix at certain quantiles}
\usage{
clamp.mat(mat, alpha = 0.01)
}
\arguments{
\item{alpha}{optional cutoff.}
}
\value{
clamped matrix
}
\description{
Clamp a matrix at certain quantiles
}
