% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{corner}
\alias{corner}
\title{Show the top corner of a matrix:}
\usage{
corner(m, x = 5)
}
\arguments{
\item{m}{matrix}

\item{x}{number of top components to show}
}
\description{
Show the top corner of a matrix:
}
