% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{cosine.dist}
\alias{cosine.dist}
\title{Cosine dist on two matrices:}
\usage{
cosine.dist(x, y = NULL, verbose = TRUE)
}
\arguments{
\item{x}{First matrix for jaccard distance}

\item{y}{Second matrix. If NULL, uses transpose of x}
}
\description{
Cosine dist on two matrices:
}
