% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_use_functions.R
\name{.ls.objects}
\alias{.ls.objects}
\title{List objects + their sizes:}
\usage{
.ls.objects(pos = 1, pattern, order.by, decreasing = FALSE,
  head = FALSE, n = 5)
}
\description{
https://stackoverflow.com/questions/1358003/
}
