% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_repel_text.R
\name{general_repel_text}
\alias{general_repel_text}
\title{Repel labels from points and each other}
\usage{
general_repel_text(x, y, labels, cex = 1, pch = 19, pt.cex = 1,
  seed = NULL, hjust = 0.5, vjust = 0.5, force = 1,
  force_pull = 1, max.iter = 2000, xlim = c(NA, NA), ylim = c(NA,
  NA), direction = "both")
}
\value{
Return new locations and facings for each label
}
\description{
Repel labels from points and each other
}
