% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{get_optimal_ordering}
\alias{get_optimal_ordering}
\title{Reorder dist. matrix according to order.optimal (cba)}
\usage{
get_optimal_ordering(distmat, method = "average")
}
\arguments{
\item{distmat}{distance matrix}

\item{method}{clustering method from hclust}
}
\description{
Reorder dist. matrix according to order.optimal (cba)
}
