% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimap_specific_functions.R
\name{go.enr}
\alias{go.enr}
\title{Compute go enrichments by over-representation test}
\usage{
go.enr(x, ont = "BP", tag = NULL, allx = gmdf$symbol, minsize = 10,
  maxsize = 500)
}
\arguments{
\item{dt}{distance object}
}
\description{
Requires clusterProfiler
}
