% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{jacc.dist}
\alias{jacc.dist}
\title{Faster jaccard distance}
\usage{
jacc.dist(x, y = NULL, verbose = TRUE)
}
\arguments{
\item{x}{First matrix for jaccard distance}

\item{y}{Second matrix. If NULL, uses transpose of x}
}
\description{
Calculate jaccard distance (1 - nint / nunion)
}
