% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimap_specific_functions.R
\name{label.runs}
\alias{label.runs}
\title{Label runs in a column of labels}
\usage{
label.runs(factor.labels, labels, rdcol)
}
\arguments{
\item{factor.labels}{Labels as factors}

\item{labels}{Labels}

\item{rdcol}{Colors for the labels}
}
\description{
Label runs in a column of labels
}
