% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general_use_functions.R
\name{make.tform}
\alias{make.tform}
\title{Make a transformation matrix for factor data}
\usage{
make.tform(x, norm = FALSE, u = NULL)
}
\description{
Make a transformation matrix for factor data
}
