% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimap_specific_functions.R
\name{meta.image}
\alias{meta.image}
\title{Plot metadata image with appropriate colors}
\usage{
meta.image(metamat, colvals = NULL, labels = NULL, horiz = FALSE,
  cex = 1, return.mat = FALSE, useRaster = FALSE)
}
\arguments{
\item{metamat}{Metadata matrix}

\item{colvals}{List of per-covariate colors for matrix}
}
\description{
Plot metadata image with appropriate colors
}
