% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing_functions.R
\name{munge.nos}
\alias{munge.nos}
\title{Extract numbers from a string}
\usage{
munge.nos(x)
}
\arguments{
\item{x}{string to search}
}
\value{
all numbers found in the string
}
\description{
Only looks for integers with commas at the moment.
}
