% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{pivot.tomatrix}
\alias{pivot.tomatrix}
\title{Turn a dataframe with key and value into a wide matrix}
\usage{
pivot.tomatrix(df, key, value)
}
\arguments{
\item{df}{3-column df}

\item{key}{column for pivot value}

\item{value}{column for values to put in matrix}
}
\description{
NOTE: From 3-column df, where column 1 is another axis of interest
}
