% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{plot.cov}
\alias{plot.cov}
\title{Plot a covariance matrix}
\usage{
\method{plot}{cov}(mat, clamp = TRUE, palette = colryb,
  breaks = NULL, breakscol = "black", blty = 1, zlim = NULL)
}
\arguments{
\item{mat}{symmetric matrix}

\item{clamp}{whether to clamp}
}
\description{
Plot a covariance matrix
}
