% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimap_specific_functions.R
\name{plot.dt.sym}
\alias{plot.dt.sym}
\title{Plot distance symmetric matrix}
\usage{
\method{plot}{dt.sym}(dt, nbreak = NULL, txtcol = "black")
}
\arguments{
\item{dt}{distance object}
}
\description{
Plot distance symmetric matrix
}
