% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimap_specific_functions.R
\name{process.dist.mat}
\alias{process.dist.mat}
\title{Process a distance matrix for the epimap samples}
\usage{
process.dist.mat(mat)
}
\description{
Process a distance matrix for the epimap samples
}
