% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{reord}
\alias{reord}
\title{Reorder matrix according to order.optimal (cba)}
\usage{
reord(mat, measure = "eJaccard", method = "complete")
}
\arguments{
\item{mat}{matrix to reorder}

\item{measure}{distance measure from dist}

\item{method}{clustering method from hclust}
}
\description{
Reorder matrix according to order.optimal (cba)
}
