% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{run.fisher}
\alias{run.fisher}
\title{Run fisher's exact test:}
\usage{
run.fisher(y)
}
\arguments{
\item{y}{Vector of contingency table values}
}
\value{
Enrichment p-value based on fisher's exact test
}
\description{
Run fisher's exact test:
}
