% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment_functions.R
\name{run.hyper}
\alias{run.hyper}
\title{Run hypergeometric test:}
\usage{
run.hyper(y)
}
\arguments{
\item{y}{Test vector of 4 values}
}
\value{
enrichment p-value based on hypergeometric
}
\description{
Give vector with number of hits, number of draws, number of white balls(hittable), number of total balls:
}
