% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing_functions.R
\name{space.1d}
\alias{space.1d}
\title{Space labels out for 1d}
\usage{
space.1d(xx, box.pad = 0.02, lim = c(0, 1))
}
\arguments{
\item{xx}{positions of text}

\item{box.pad}{box size/padding for text}

\item{lim}{limits of space to use}
}
\description{
Space labels out for 1d
}
