% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/text_processing_functions.R
\name{split.text}
\alias{split.text}
\title{Split text by widths}
\usage{
\method{split}{text}(x, width = 90)
}
\arguments{
\item{x}{string to split}

\item{x}{text width}
}
\description{
Split text by widths
}
