% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix_functions.R
\name{tsp.col}
\alias{tsp.col}
\title{Adust color transparency for hex color}
\usage{
tsp.col(x, alpha = 0.5)
}
\arguments{
\item{x}{color}

\item{alpha}{transparency level}
}
\value{
adjusted color
}
\description{
Adust color transparency for hex color
}
