% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/epimap_specific_functions.R
\name{UpSet.v2}
\alias{UpSet.v2}
\title{Upset plot (V2 for coloring bars) for set intersections}
\usage{
UpSet.v2(m, comb_col = "black", set_col = "black", lwd = 2,
  pt_size = unit(3, "mm"), bg_col = "#F0F0F0", bg_pt_col = "#CCCCCC",
  set_order = order(set_size(m), decreasing = TRUE), comb_order = if
  (attr(m, "set_on_rows")) {     order.comb_mat(m[set_order, ], decreasing
  = TRUE) } else {     order.comb_mat(m[, set_order], decreasing = TRUE) },
  top_annotation = upset_top_annotation(m),
  right_annotation = upset_right_annotation(m),
  row_names_side = "left", ...)
}
\description{
Note: Requires ComplexHeatmap
}
